Add dynamic color functionality, update styles
Add dynamic color functionality to ProjectListView.

Add get_context_data and get_colors methods to generate a list of similar colors to be sent to the template for list item backgrounds. The list of colors is the same length as the queryset.

Add styled header.

Update styles in listing area. Add functionality from color context variable.