Fix user resources list error
Update UserLearningResourceListView to include logic that will prevent the error that occurs when a user without any LearningResource objects tries to access the page that uses this view.

Update page template learning_resources_list to include a helpful message if the user is trying to access this page before adding any  of their own LearningResource objects.